Case D - page 2 of CA inclusion in code
To account for case where Type 2 is wrong but Type 3 and Type 4 are right i.e. which as discussed in class as Case D at the bottom of page 2 of the CA ; Case D would mean that the prediction accuracy of the model should evaluate to 0% (versus 67% for Type 3 and Type 4 being right i.e. this 67% is not possible if the base Type 2 is wrong i.e. resulting in overall prediction being 0% in accuracy)